src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "course project"
authors = [
    {name = "Ekaterina Ambartsumova",email = "ekaterina.amba@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy[asyncio] (>=2.0.43,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "asyncpg (>=0.30.0,<0.31.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
mypy = "^1.17.1"
ruff = "^0.12.9"
pytest = "^8.4.1"

